Allow deletion of unmerged branches
Calling git branch with the `-d` flag will only delete a branch if it
has been fully merged with its parent (usually master). In this case, if
the user explicitly chooses to delete a local branch, we want to honour
this request regardless of whether the branch has been merged.

This commit swaps `-d` with `-D`, which will proceed with the deletion
regardless of merge status.